[SFTool] Move Testing Config into Repo
Summary:

This diff makes several chagnes to sftool so that the testing configuration (schemes, simulator & os, etc.) can be stored in vcs.
This makes running the tests in the C/I environment easier, and allows the us to leave make less changes to the c/i config since it will read the latest configuration when the changes are pulled.

- Rename the config file to `.sftool-config`
- Add fields in the config file to store simulator device name and OS
- Add the `test` command to invoke `xctest` with the appropriate configuration
- Also add the ability to read the currently installed analytics configuration

Test Plan: N/a

Reviewers:

Subscribers: